Moved login and signup forms into their own component functions.
